Moving datasource to its own package
